dev(block_import_refactoring): Moved types to primitives
