+++ crates/primitives/class/Cargo.toml
@@ -37,6 +37,7 @@ flate2 = { workspace = true }
 num-bigint = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+rayon = { workspace = true }
 
 [dev-dependencies]
 starknet-providers = { workspace = true }
